refactor: move ItemThumbnail to separate directory
